DEP: Remove Scripts, Resources, Tests, Sample_Code (#845)
* All of them are removed from the package distributions
* Scripts is additionally moved to the cpdf project
* Sample_Code is moved to the docs
