--- Sample_Code/hierarchical_toc.md
@@ -1,3 +1,0 @@
-# Example of hierarchical TOC from several pdf files
-
-https://github.com/vb64/pdftoc
