+++ resources/crazyones.txt
