+++ resources/imagemagick-ASCII85Decode.pdf
