+++ resources/imagemagick-CCITTFaxDecode.pdf
