+++ resources/jpeg.txt
