+++ resources/side-by-side-subfig.pdf
