+++ tests/test_merger.py
@@ -1,44 +1,17 @@
 import os
+import sys
 
 import PyPDF2
-from PyPDF2 import PdfFileReader
 from PyPDF2.generic import Destination
 
 TESTS_ROOT = os.path.abspath(os.path.dirname(__file__))
 PROJECT_ROOT = os.path.dirname(TESTS_ROOT)
-RESOURCE_ROOT = os.path.join(PROJECT_ROOT, "Resources")
+RESOURCE_ROOT = os.path.join(PROJECT_ROOT, "resources")
+
+sys.path.append(PROJECT_ROOT)
 
 
-def page_ops(pdf_path, password):
-    pdf_path = os.path.join(RESOURCE_ROOT, pdf_path)
-
-    reader = PdfFileReader(pdf_path)
-
-    if password:
-        reader.decrypt(password)
-
-    page = reader.pages[0]
-    page.mergeRotatedScaledPage(page, 90, 1, 1)
-    page.mergeScaledTranslatedPage(page, 1, 1, 1)
-    page.mergeRotatedScaledTranslatedPage(page, 90, 1, 1, 1, 1)
-    page.addTransformation([1, 0, 0, 0, 0, 0])
-    page.scale(2, 2)
-    page.scaleBy(0.5)
-    page.scaleTo(100, 100)
-    page.compressContentStreams()
-    page.extractText()
-
-
-def test_page_operations(benchmark):
-    """
-    Apply various page operations.
-
-    Rotation, scaling, translation, content stream compression, text extraction
-    """
-    benchmark(page_ops, "libreoffice-writer-password.pdf", "openpassword")
-
-
-def merge():
+def test_merge():
     pdf_path = os.path.join(RESOURCE_ROOT, "crazyones.pdf")
     outline = os.path.join(RESOURCE_ROOT, "pdflatex-outline.pdf")
     pdf_forms = os.path.join(RESOURCE_ROOT, "pdflatex-forms.pdf")
@@ -93,12 +66,3 @@ def merge():
 
     # Clean up
     os.remove(tmp_path)
-
-
-def test_merge(benchmark):
-    """
-    Apply various page operations.
-
-    Rotation, scaling, translation, content stream compression, text extraction
-    """
-    benchmark(merge)
