+++ tox.ini
@@ -7,4 +7,4 @@ deps =
 	pillow
 	pytest
 	pytest-cov
-commands = pytest Tests --cov --cov-report term-missing -vv
+commands = pytest tests --cov --cov-report term-missing -vv
